requirements.txt
accelerate==1.8.1
aiofiles==23.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.13.0
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
boto3==1.39.1
botocore==1.39.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
cryptography==45.0.4
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.104.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
huggingface-hub==0.33.0
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
openai==1.86.0
openpyxl==3.1.2
packaging==25.0
pandas==2.1.4
parso==0.8.4
passlib==1.7.4
peft==0.16.0
pillow==10.4.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pydub==0.25.1
Pygments==2.19.2
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytube==15.0.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
sympy==1.14.0
tokenizers==0.21.2
torch==2.2.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
yt-dlp==2025.6.9
